+++ .cz-config.js
@@ -2,7 +2,7 @@
  * @Author: cpasion-office-win10 373704015@qq.com
  * @Date: 2024-02-19 09:18:09
  * @LastEditors: cpasion-office-win10 373704015@qq.com
- * @LastEditTime: 2024-02-20 15:32:58
+ * @LastEditTime: 2024-02-20 16:07:11
  * @FilePath: \cps-blog-docusaurus-v3\.cz-config.js
  * @Description: 这是默认设置,请设置`customMade`, 打开koroFileHeader查看配置 进行设置: https://github.com/OBKoro1/koro1FileHeader/wiki/%E9%85%8D%E7%BD%AE
  */
@@ -20,36 +20,26 @@
 
 module.exports = {
   types: [
-    { value: "✨新增", name: "✨:art:新增:    新增功能或者特性" },
-    { value: "⚡️性能", name: "⚡️性能:    提升性能" },
-    { value: "🔃更新", name: "🔃更新:    更新一些内容" },
-    { value: "✅测试", name: "✅测试:    添加一个测试" },
-    { value: "🐛修复", name: "🐛修复:    修复一个Bug" },
-    { value: "📝文档", name: "📝文档:    变更的只有文档" },
-    { value: "💄格式", name: "💄格式:    空格, 分号等格式修复" },
-    { value: "♻️重构", name: "♻️重构:    代码重构，注意和特性、修复区分开" },
-    { value: "🔧工具", name: "🔧工具:    开发工具变动(构建、脚手架工具等)" },
-    { value: "⏪回滚", name: "⏪回滚:    代码回退" },
+    { value: "✨feat", name: "新增:    新增功能或者特性" },
+    { value: "⚡️perf", name: "优化:    优化相关，比如提升性能、体验" },
+    { value: "💄style", name: "格式:    💄仅仅修改了空格、格式缩进等等，不改变代码逻辑" },
+    { value: "✅test", name: "测试:    测试用例，包括增加缺失用例或者修正测试用例" },
+    { value: "🐛fix", name: "修复:    修复Bug" },
+    { value: "📝docs", name: "文档:    仅仅修改了文档，比如`README`, `CHANGELOG`, CONTRIBUTE等等" },
+    { value: "♻️refactor", name: "重构:    重构（即不是新增功能，也不是修改bug的代码变动）" },
+    { value: "🔧chore", name: "工具:    构建过程或辅助工具的变动" },
+    { value: "⏪revert", name: "回滚:    用于撤销以前的 commit，后面跟着被撤销 Commit 的 Header。" },
+    { value: "🔃ci", name: "更新:    更改我们的CI配置文件和脚本" },
   ],
-  // scopes: [
-  //   { name: 'leetcode' },
-  //   { name: 'javascript' },
-  //   { name: 'typescript' },
-  //   { name: 'Vue' },
-  //   { name: 'node' },
-  // ],
+  // scopes: [{ name: "leetcode" }, { name: "javascript" }, { name: "typescript" }, { name: "Vue" }, { name: "node" }],
   // it needs to match the value for field type. Eg.: 'fix'
-  /*  scopeOverrides: {
-      fix: [
-        {name: 'merge'},
-        {name: 'style'},
-        {name: 'e2eTest'},
-        {name: 'unitTest'}
-      ]
-    },  */
+  scopeOverrides: {
+    fix: [{ name: "merge" }, { name: "style" }, { name: "e2eTest" }, { name: "unitTest" }],
+  },
   // override the messages, defaults are as follows
   messages: {
     type: "选择一种你的提交类型:",
+    scopes: "影响范围",
     scope: "选择一个scope (可选):",
     // used if allowCustomScopes is true
     customScope: "Denote the SCOPE of this change:",