+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * ✨新增(custom): 添加changlog自动生成 ([c4c9775](https://github.com/mucpsing/blog-docusaurus-v3/commit/c4c9775))
 * ✨新增(custom): git cz ([f7a763b](https://github.com/mucpsing/blog-docusaurus-v3/commit/f7a763b))
 * 🔃更新(custom): 优化package.json/scripts的命令，添加npm run git ([7a261fd](https://github.com/mucpsing/blog-docusaurus-v3/commit/7a261fd))
+* 🔧chore(custom): 完善git cz的说明 ([230b70e](https://github.com/mucpsing/blog-docusaurus-v3/commit/230b70e))
 * chore: init ([8d3a70a](https://github.com/mucpsing/blog-docusaurus-v3/commit/8d3a70a))
 * chore: init ([b9e359c](https://github.com/mucpsing/blog-docusaurus-v3/commit/b9e359c))